Moved `customPokemonData` into types folder + fixed comments
