+++ src/data/mystery-encounters/utils/encounter-phase-utils.ts
@@ -49,7 +49,7 @@ import { TrainerSlot } from "#enums/trainer-slot";
 import type PokemonSpecies from "#app/data/pokemon-species";
 import type { IEggOptions } from "#app/data/egg";
 import { Egg } from "#app/data/egg";
-import type { CustomPokemonData } from "#app/data/custom-pokemon-data";
+import type { CustomPokemonData } from "#app/@types/pokemon-data";
 import type HeldModifierConfig from "#app/interfaces/held-modifier-config";
 import { MovePhase } from "#app/phases/move-phase";
 import { EggLapsePhase } from "#app/phases/egg-lapse-phase";
