+++ src/system/version_migration/versions/v1_0_4.ts
@@ -2,7 +2,7 @@ import { SettingKeys } from "#app/system/settings/settings";
 import type { SystemSaveData, SessionSaveData } from "#app/system/game-data";
 import { AbilityAttr, defaultStarterSpecies, DexAttr } from "#app/system/game-data";
 import { allSpecies } from "#app/data/pokemon-species";
-import { CustomPokemonData } from "#app/data/custom-pokemon-data";
+import { CustomPokemonData } from "#app/@types/pokemon-data";
 import { isNullOrUndefined } from "#app/utils/common";
 import type { SystemSaveMigrator } from "#app/@types/SystemSaveMigrator";
 import type { SettingsSaveMigrator } from "#app/@types/SettingsSaveMigrator";
